+++ precode.go
@@ -2,10 +2,8 @@ package main
 
 import (
     "net/http"
-    "net/http/httptest"
     "strconv"
     "strings"
-    "testing"
 )
 
 var cafeList = map[string][]string{
@@ -46,13 +44,3 @@ func mainHandle(w http.ResponseWriter, req *http.Request) {
     w.Write([]byte(answer))
 }
 
-func TestMainHandlerWhenCountMoreThanTotal(t *testing.T) {
-    totalCount := 4
-    req := ... // здесь нужно создать запрос к сервису
-
-    responseRecorder := httptest.NewRecorder()
-    handler := http.HandlerFunc(mainHandle)
-    handler.ServeHTTP(responseRecorder, req)
-
-    // здесь нужно добавить необходимые проверки
-}
